Fix: deg_to_rad change to doubl

+++ utils/helpers.c
@@ -6,13 +6,13 @@
 /*   By: orezek <orezek@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/17 18:40:42 by orezek            #+#    #+#             */
-/*   Updated: 2024/06/22 19:19:02 by orezek           ###   ########.fr       */
+/*   Updated: 2024/06/26 12:51:39 by orezek           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../cube.h"
 
-float	deg_to_rad(int a)
+double	deg_to_rad(int a)
 {
 	return (a * M_PI / 180.0);
 }
